New service loading procedure
