+++ smallcloud.js
@@ -1,11 +1,20 @@
 
+var init = require('./core');
 
-var smallcloud = require('./core');
-var init       = require('./core/lib/init');
+// XXX
+console.log('>>>', init);
 
 
-// Setup core object
-var S = smallcloud();
 
-// Initialize smallcloud
-init(S);
+
+
+
+//var smallcloud = require('./core');
+//var init       = require('./core/lib/init');
+//
+//
+//// Setup core object
+//var S = smallcloud();
+//
+//// Initialize smallcloud
+//init(S);
